+++ pkg/symbolizer/symbolizer_test.go
@@ -479,7 +479,6 @@ func setup(t *testing.T) (*grpc.ClientConn, pb.MetastoreServiceClient, *Symboliz
 		metastore,
 		table,
 		schema,
-		false,
 	)
 
 	lis, err := net.Listen("tcp", "127.0.0.1:0")