+++ ui/packages/shared/components/src/ParcaContext/index.tsx
@@ -16,9 +16,9 @@ import Spinner from '../Spinner';
 
 interface Props {
   loader: ReactNode;
-  perf: {
-    onRender: ProfilerOnRenderCallback;
-    markInteraction: (interactionName: string) => void;
+  perf?: {
+    onRender?: ProfilerOnRenderCallback;
+    markInteraction: (interactionName: string, sampleCount: number | string) => void;
   };
 }