+++ ui/packages/shared/profile/src/ProfileView/index.tsx
@@ -240,7 +240,10 @@ export const ProfileView = ({
               <div ref={ref} className="flex space-x-4 justify-between w-full">
                 {dashboardItems.includes('icicle') && flamegraphData?.data != null && (
                   <div className="w-full">
-                    <Profiler id="icicleGraph" onRender={perf.onRender}>
+                    <Profiler
+                      id="icicleGraph"
+                      onRender={perf?.onRender as React.ProfilerOnRenderCallback}
+                    >
                       <ProfileIcicleGraph
                         curPath={curPath}
                         setNewCurPath={setNewCurPath}