+++ ui/packages/app/grafana-panel-plugin/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.27](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.26...@parca/grafana-panel@0.0.27) (2022-11-29)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## [0.0.26](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.25...@parca/grafana-panel@0.0.26) (2022-11-29)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## 0.0.25 (2022-11-29)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## [0.0.24](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.23...@parca/grafana-panel@0.0.24) (2022-11-24)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## [0.0.23](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.22...@parca/grafana-panel@0.0.23) (2022-11-22)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## [0.0.22](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.21...@parca/grafana-panel@0.0.22) (2022-11-22)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## 0.0.21 (2022-11-21)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## [0.0.20](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.19...@parca/grafana-panel@0.0.20) (2022-11-17)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## 0.0.19 (2022-11-17)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## 0.0.18 (2022-11-16)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
+## [0.0.17](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.16...@parca/grafana-panel@0.0.17) (2022-11-15)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
 ## 0.0.16 (2022-11-14)
 
 **Note:** Version bump only for package @parca/grafana-panel